refactoring: move business logic from model to separate service class, and keep in model only database related code. added exception throwing if errors occurred
