+++ app/Exceptions/FileUploadException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types = 1);
+
+namespace App\Exceptions;
+
+class FileUploadException extends \Exception
+{
+}
