+++ app/Helpers/TransactionFormatter.php
@@ -4,13 +4,13 @@ namespace App\Helpers;
 
 class TransactionFormatter
 {
-    public static function dollars($amount): string
+    public static function dollars(string $amount): string
     {
         $isNegative = $amount < 0;
         return ($isNegative ? '-' : '') . '$' . number_format(abs($amount), 2);
     }
 
-    public static function date($date)
+    public static function date(string $date): string
     {
         return \DateTime::createFromFormat('Y-m-d', $date)->format('F j, o');
     }
